Research, Design Doc & moodboards
I have added research of Boss fights and combat systems that could potentially aid our game in terms of flow and balance of game. I have created the Design Document which is not finished but I have also uploaded that into Evernote, so that we can all have a shared version of it that can be edited by any member of the group. I have also started creating moodboards for the artsytle of our game.