AI & A* Pathfinding
Added the A* Pathfinding Project to the game, it allows our enemies to move around the map using A* without writing our own A* pathfinding. Free version is permitted to be used in commercial projects.

TestAI will follow the plauer around the map, avoiding any obstacles in the way. They will look at the player, and they will shoot the player when the player is in sight. If the player is out of sight, they will chase the player.

DummyAI - just a dummy.