Melee Character + Improvements to AI
Added a melee character, which uses the mouse position to aim his "sword". At the moment there is no swing animation, but the damage is dealt.

AI now properly uses health, and will die when the health is depleted.

A* now fully works, the Enemy Controller takes all the parameters and passes them to the A* script; this way it is easier for everyone to edit the enemies.

Player can now die, and thanks to the A* script if player is dead, the console is not spammed with errors. This will make things easier in the future (with respawning, etc.)