Updated Boss & Bullet
Created parent scripts for the bullet. At the moment there are two types of bullets, BossBullets and PlayerBullets. Each specific bullet will inherit from their respective parent. RangedAI at the moment is using player bullet.

BossBattle is finished, apart from the thump attack, as it needs to push the player away, create some sort of dust particles upon impact and damage the player if they are in the range of the attack.

Level Gen seems to be detecting a boss room, but it doesn't seem to spawn the boss when I want to. I remove the functionality for now and will add it back once its working as intended.