made a new unity file
This is so my group can see any work i have done that is not necessarily on the game project unity file but its my experiments on the modular assets. I have decided to make it on here so it does not disrupt any part of our actual game. I have also made a placeholder for the tavern (character select) so that players can get the feel for the game during playtesting, as i make the actual assets for the character select