Shop Spwning + Boss Icon
Shop will now spawn a minimap icon to indicate which room the shop is in, next I will add the actual shop spawning in the room.

Boss icon now also spawns on the minimap, to indicate which room the boss portal spawns in.

Added minimap to the normal rooms; all rooms now have minimap pictures displayed on the minimap.

Added minimap tiles to the entrance rooms; all entrance rooms now have minimap tiles displayed on the minimap.

Renamed Room generation, to Level Generation as it better represents the fucntionality.

Working on the new character stats system, which will allow to add/remove modifiers to player's stats.