Loot Room Logic
- Added loot room logic.

Loot Rooms:
When players enter the room, they can open a chest which will yield a random item from drop pool. When triggered, enemies will spawn in the room, and will attack the player while the doors lock the player in the room. Only way out is to kill all the enemies.